PickUp baru setengah
